URL Encoding on fields list.
If someone adds a field with a hash char in it, it will break the
fields= url, as it won't be encoded properly and will cause all fields
after the one with a hash in to be ignored in the response.
This adds encodeurl() around the fields= list so they should all be
loaded properly.